feat: Add CustomTooltip and HotelLegend components; refactor AnalyseConcurentiellePage for improved hotel visibility and date handling
